adjust size type in itemcontainer to accommodate larger inventories

+++ src/servers/sapphire_zone/Inventory/ItemContainer.cpp
@@ -11,7 +11,7 @@
 
 extern Sapphire::Framework g_fw;
 
-Sapphire::ItemContainer::ItemContainer( uint16_t storageId, uint8_t maxSize, const std::string& tableName,
+Sapphire::ItemContainer::ItemContainer( uint16_t storageId, uint16_t maxSize, const std::string& tableName,
                                     bool isMultiStorage, bool isPersistentStorage ) :
   m_id( storageId ),
   m_size( maxSize ),
@@ -32,12 +32,12 @@ uint16_t Sapphire::ItemContainer::getId() const
   return m_id;
 }
 
-uint8_t Sapphire::ItemContainer::getEntryCount() const
+uint16_t Sapphire::ItemContainer::getEntryCount() const
 {
-  return static_cast< uint8_t >( m_itemMap.size() );
+  return static_cast< uint16_t >( m_itemMap.size() );
 }
 
-void Sapphire::ItemContainer::removeItem( uint8_t slotId )
+void Sapphire::ItemContainer::removeItem( uint16_t slotId )
 {
   auto pLog = g_fw.get< Logger >();
   auto pDb = g_fw.get< Db::DbWorkerPool< Db::ZoneDbConnection > >();
@@ -68,9 +68,9 @@ const Sapphire::ItemMap& Sapphire::ItemContainer::getItemMap() const
   return m_itemMap;
 }
 
-int8_t Sapphire::ItemContainer::getFreeSlot()
+int16_t Sapphire::ItemContainer::getFreeSlot()
 {
-  for( uint8_t slotId = 0; slotId < m_size; slotId++ )
+  for( uint16_t slotId = 0; slotId < m_size; slotId++ )
   {
     ItemMap::iterator it = m_itemMap.find( slotId );
     if( it == m_itemMap.end() ||
@@ -80,7 +80,7 @@ int8_t Sapphire::ItemContainer::getFreeSlot()
   return -1;
 }
 
-Sapphire::ItemPtr Sapphire::ItemContainer::getItem( uint8_t slotId )
+Sapphire::ItemPtr Sapphire::ItemContainer::getItem( uint16_t slotId )
 {
 
   if( ( slotId > m_size ) )
@@ -93,7 +93,7 @@ Sapphire::ItemPtr Sapphire::ItemContainer::getItem( uint8_t slotId )
   return m_itemMap[ slotId ];
 }
 
-void Sapphire::ItemContainer::setItem( uint8_t slotId, ItemPtr pItem )
+void Sapphire::ItemContainer::setItem( uint16_t slotId, ItemPtr pItem )
 {
   if( slotId > m_size )
     return;
@@ -101,7 +101,7 @@ void Sapphire::ItemContainer::setItem( uint8_t slotId, ItemPtr pItem )
   m_itemMap[ slotId ] = pItem;
 }
 
-uint8_t Sapphire::ItemContainer::getMaxSize() const
+uint16_t Sapphire::ItemContainer::getMaxSize() const
 {
   return m_size;
 }
